refactor(models): remove one level for places
Place will be directly linked to Education now, so EducationPlace is not needed.
In an ideal world, a shared list of places might be a good idea, but it is probably
forcing the users too much. In the end every stagebureau works with  their own
list of places so it might be easier like this.