refactor: merge ImportantDate and CustomForm
Previous implementation linked the anwsers to a form directly, but that
made impossible to easily limit answers for a period of time. Forms were
reusable, but then again, we want to be able to link answers to a specific
questioning period.

This commit merges the two models, and moves the form definition to the
Questioning model.