feat: simple xss escape for custom form data
Django querysets are normally escaped, but we now have extra
html escaping in custom forms, just in case. These are open to
different users so why not.

In any case, it should be noted that all information displayed on Metis
uses Django or Vue templates, and these are protected by default:

- https://docs.djangoproject.com/en/dev/topics/security/
- https://vuejs.org/guide/best-practices/security.html