refactor: use S3 as file storage and add a "file guard" for acl
All files are stored privately on S3, and then Django takes care of showing
the file (or not) based on user permissions and profile. To access S3 files,
a temporary presigned url is generated (for 10 seconds) and the contents
are then streamed to the client, so the S3 url is never shared.