refactor: simplify invitation process
Now that we only use UGent login, we create Metis accounts already when
inviting contacts. This makes it clearer for the stagebureau as we can directly
show the created accounts on their lists, and it gives users the right access
rights as soon as they create a UGent account, without the need of using a
personal invitation link.

The invitation email can show better information about the UGent login
process (Microsoft accounts, etc,) and just point to the Metis homepage or
user dashboard. The link will never expire and they will always have a valid
entry point in their emails.